Update SCCS metadata mapping and fix FinalizeReport (#170)
* Update SCCS metadata mapping JSON file

Why these changes are being introduced:
* The metadata mapping JSON file required the addition of a few fields
in order to fulfill a batch deposit request. It was decided to take
the approach of updating the metadata mapping JSON file as needed
rather than reworking the metadata mapping process at this stage.

How this addresses that need:
* Add 'dc.contributor.department' and 'dc.description.abstract' as fields
* Replace 'nan' with None when reading records from metadata CSV file
* Fix counts in FinalizeReport
* Update SimpleCSV to exclude DSpace metadata JSON files retrieving bitstream uris

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/INFRA-509