Update DSC workflows to use 'S3_BUCKET' env var (#171)
Why these changes are being introduced:
* The S3 bucket configured for DSC will include MITL's
AWS account ID to ensure that the name of the bucket meets
S3's requirement of uniqueness "across all AWS accounts in all
the AWS regions within a partition". Since the AWS account IDs
differ for Dev, Stage, and Prod, it would be difficult to dyna-
mically generate the name of the S3 bucket; it also wouldn't be
ideal to include the AWS account IDs as part of version control.
Accessing the name of the S3 bucket via environment variable
is the ideal and most secure option.

How this addresses that need:
* Add 'S3_BUCKET' as a required env var
* Set Workflow.s3_bucket as a final, property method that accesses
'S3_BUCKET' env var via Config
* Remove 's3_bucket' property methods from all workflows

Side effects of this change:
* The environment variable 'S3_BUCKET' must be set as part of the
ECS task definition.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1227