Rename environment variables
Why these changes are being introduced:
* Originally, the plan was to rename only the 'S3_BUCKET' env var,
but this prompted a larger discussion regarding naming conventions
for app env vars. The renames are intended to include descriptive
information in the name, telling the user of its intended use.
All variables relate to the app (DSC) unless otherwise specified.
Service names are included in the name only if the env var is tightly
coupled to services (e.g., SQS)

How this addresses that need:
* Rename env vars for S3 bucket, source email, and DSS input SQS queue

Side effects of this change:
* Infra must be updated to reference the new env var names

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1236