Create model for item submission DynamoDB table (#179)
Why these changes are being introduced:
* To establish idempotency for DSC workflows, it was
decided that a DynamoDB table would be used to track the
state of an "item" during workflow executions.
The dsc.db.item module defines a PynamoDB model,
which serves as a Pythonic interface to interact with
the proposed DynamoDB table.

How this addresses that need:
* Define 'ItemSubmissionDB' model using pynamodb.models.Model
* Add 'ITEM_TABLE_NAME' as required env var
* Create custom exception for PutError caused by 'ConditionalCheckFailedException'

Side effects of this change:
* DSC CLI commands will need to be updated to include
required read/write operations to DynamoDB table.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1315