Update 'finalize' command to use DynamoDB table
Why these changes are being introduced:
* This work is an important part of ensuring that running
DSC is idempotent via tracking the state of item submissions
using a DynamoDB table. With 'finalize' being the last step
of the DSC workflow, it is during this step that the table
is updated with information regarding the attempted deposit
into DSpace and whether the item submission was ingested.

How this addresses that need:
* Break up now-deprecated result message parsing method into two
methods for 'MessageAttributes' and 'Body' content
  - Each method raises an 'InvalidSQSMessageError' exception if invalid
* Rework logic in Workflow.process_sqs_queue to use new parsing methods
* Include calls to update records in DynamoDB table
* Update language in 'sqs_results_summary' count variable

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1318