Remove Demo workflow from app
Why these changes are being introduced:
* The Demo workflow was always intended to be temporary and was
initially created for testing parts of DSC with 'wiley-deposits'
infrastructure. Now that DSC infrastructure deployed in Dev
(and soon to be deployed in Stage and Prod), the Demo workflow
can be deprecated.

How this addresses that need:
* Remove metadata mapping JSON file for Demo workflow
* Remove Demo workflow module

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1167