Rework ItemSubmission to better encapsulate data persistence layer
Why these changes are being introduced:
* Initial work to update the 'reconcile' and 'submit' commands to use the refactored
ItemSubmission class underscored the need for additional updates to improve the
ergonomics of the class and emphasize the encapsulation of data persistance
as part of the domain model. While most of the code was already in place,
these updates focus on some renaming and slight reorganization of methods related
to database operations. The renames were intended to reflect that pulling data
from the DynamoDB table is an inherent part of creating an ItemSubmission.

How this addresses that need:
* Move logic for loading data into ItemSubmission into a private method ('_from_db')
* Rename 'update_db' -> 'upsert_db' to clarify insertion of new records
* Rework existing methods to get, get_batch, and create methods
* Update Workflow.reconcile_items to use get and create methods
* Update Workflow.submit_items to use get_batch to iterate over ItemSubmission
instances for a given batch

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1355