Align workflow methods for iterating over item metadata (#195)
Why these changes are being introduced:
* For the OpenCourseWare (OCW) workflow, the method would previously raise
a FileNotFoundError if a metadata JSON file (data.json) was not found
in the zip file. These changes ensure that OCW returns a dict
regardless (containing at least the item identifier). This was identified
as a prerequisite to also aligning the 'reconcile' pattern with
the other DSC CLI commands.

How this addresses that need:
* Add try-except block to OpenCourseWare.item_metadata_iter
* Add unit test to validate OpenCourseWare.metadata_mapping
* Fix: Update OCW metadata mapping JSON file

Side effects of this change:
* This also fixes a breaking change that was introduced by the
previous update that introduced a metadata transformer for OCW.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1431