Include source metadata as a processing attribute of ItemSubmission
Why these changes are being introduced:
This change is in line ensuring all key processes are performed with the
ItemSubmission class. This is a pre-requisite for defining a record-level
'reconcile_item' abstract method that uses the ItemSubmission class.

How this addresses that need:
* Add 'source_metadata' attribute to ItemSubmission
* Set 'source_metadata' in Workflow.reconcile_items

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1433