Cache list of bitstream URIs to Workflow class (#196)
Why these changes are being introduced:
Every time the S3Client.files_iter method is called, it sends an API
request to list the contents of an S3 bucket. This ensures that the list
of bitstreams is cached to the Workflow, the first time it is referenced
in the code and updates all dependent methods to use the cache.

How this addresses that need:
* Add property method for retrieving cached bitstream URIs to base Workflow
* Add abstract method to get bitstream URIs to base workflow
* Update OpenCourseWare and SimpleCSV workflows to use cached bitstream URIs
* Replace get_bitstream_s3_uris with final method (get_item_bitstream_uris)

Side effects of this change:
* Reduce number of requests to access Amazon S3 services.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1432