Update batch creation report to capture errors (#205)
Why these changes are being introduced:
* While errors during the 'submit' and 'finalize' steps
are recorded in the DynamoDB table, errors that occur during
batch creation (the 'create' step) need to be captured and
included in reporting through some method. The purpose of these
changes are to pass along the list of errors returned by
Workflow.prepare_batch to the reporting module, CreateReport,
which renders a CSV file to report on item submissions
that fail batch creation.

How this addresses that need:
* Add 'errors' instance attr to CreateReport
* Define method for writing an 'errors' CSV file on CreateReport
* Define method for generating attachments
* Update signature of Workflow.send_report to accept instantiated
Reports (instead of classes)
* Add try-except block in 'create' CLI command to handle
'BatchCreationFailedError'

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1512